Remove and reoganize around new methods
Realized what I was trying to do wouldn't work, thought about it some and decided to try again with something new. New plan is to
foreach the videos list and find the first thing that qualifies as being the primary for *any* episode, then remove it and pipe the remaining
videos into a function that will find any alternate versions for that episode and add them to the VideoInfo file, before returning it
to be added as an entry in a new consolidatedEpisodes list. This continues until the original videos list is empty, then I hope
to group/sort and return it?